+++ CHANGES.md
@@ -8,6 +8,7 @@
 	- `F [character]`: Jump to the previous entry starting with `[character]` (repeatable)
 	- `;`: Jump to the next entry matching the most recent find command (repeatable)
 	- `,`: Jump to the previous entry matching the most recent find command (repeatable)
+- Fixed hanging when opening files when your default editor takes awhile to open (looking at you, Visual Studio)
 
 ## 0.3.1
 - Added `g` (top of list) and `G` bottom of list motions ([#5](https://github.com/LPGhatguy/magic-school-bus/issues/5))